+++ tests/compile_fail/with_settings/bump/borrow_mut/guaranteed_allocated_decrease.rs
@@ -1,0 +1,19 @@
+use bump_scope::{
+    Bump,
+    alloc::Global,
+    settings::{BumpAllocatorSettings, BumpSettings},
+};
+
+type In = <BumpSettings as BumpAllocatorSettings>::WithGuaranteedAllocated<true>;
+type Out = <BumpSettings as BumpAllocatorSettings>::WithGuaranteedAllocated<false>;
+
+fn convert(bump: &mut Bump<Global, In>) -> &mut Bump<Global, Out> {
+    bump.borrow_mut_with_settings()
+}
+
+fn main() {
+    let mut input = Bump::<Global, In>::new();
+    let output = convert(&mut input);
+    let test = output.alloc_str("test");
+    println!("{test}");
+}