+++ tests/compile_fail/with_settings/bump_scope/borrow/guaranteed_allocated_increase.rs
@@ -1,0 +1,19 @@
+use bump_scope::{
+    Bump, BumpScope,
+    alloc::Global,
+    settings::{BumpAllocatorSettings, BumpSettings},
+};
+
+type In = <BumpSettings as BumpAllocatorSettings>::WithGuaranteedAllocated<false>;
+type Out = <BumpSettings as BumpAllocatorSettings>::WithGuaranteedAllocated<true>;
+
+fn convert<'a, 'b>(bump: &'b BumpScope<'a, Global, In>) -> &'b BumpScope<'a, Global, Out> {
+    bump.borrow_with_settings()
+}
+
+fn main() {
+    let input = Bump::<Global, In>::new();
+    let output = convert(input.as_scope());
+    let test = output.alloc_str("test");
+    println!("{test}");
+}