+++ tests/compile_fail/with_settings/bump_scope/borrow_mut/minimum_alignment_decrease.rs
@@ -1,0 +1,21 @@
+use bump_scope::{
+    Bump, BumpScope,
+    alloc::Global,
+    settings::{BumpAllocatorSettings, BumpSettings},
+};
+
+type In = <BumpSettings as BumpAllocatorSettings>::WithMinimumAlignment<2>;
+type Out = <BumpSettings as BumpAllocatorSettings>::WithMinimumAlignment<1>;
+
+fn convert<'a, 'b>(bump: &'b mut BumpScope<'a, Global, In>) -> &'b mut BumpScope<'a, Global, Out> {
+    bump.borrow_mut_with_settings()
+}
+
+fn main() {
+    let mut input = Bump::<Global, In>::new();
+    let output = convert(input.as_mut_scope());
+    let test = output.alloc_str("test");
+    println!("{test}");
+}
+
+