+++ tests/compile_fail/with_settings/bump_scope/borrow_mut/up_to_true.rs
@@ -1,0 +1,19 @@
+use bump_scope::{
+    Bump, BumpScope,
+    alloc::Global,
+    settings::{BumpAllocatorSettings, BumpSettings},
+};
+
+type In = <BumpSettings as BumpAllocatorSettings>::WithUp<false>;
+type Out = <BumpSettings as BumpAllocatorSettings>::WithUp<true>;
+
+fn convert<'a, 'b>(bump: &'b mut BumpScope<'a, Global, In>) -> &'b mut BumpScope<'a, Global, Out> {
+    bump.borrow_mut_with_settings()
+}
+
+fn main() {
+    let mut input = Bump::<Global, In>::new();
+    let output = convert(input.as_mut_scope());
+    let test = output.alloc_str("test");
+    println!("{test}");
+}